Add Cow class, methods to place cows, graphics
Add the Cow class, which has a position and a graphic, not unlike
Flower's implementation. In fact, it's almost exactly the same
implementation, they should be merged.

layer1 is now layerCows, containing the cows in a separate layer.
Cows can be added with Ranch.addCow() or Ranch.addCows().

The three cow graphics are assets/cow/din.png, assets/cow/normal.png,
and assets/cow/surf.png.