Add Interval class, will be used for domain/range
Interval represents an interval of numbers; i.e. a range with a
start and end, inclusive and exclusive.

Interval.Bound is simply a data class with a number and either
inclusive or exclusive (a boolean, inclusive). It has lessThan,
greaterThan, compareTo methods.

Interval.EqualBoundsException is thrown when equal bounds are passed
to the Interval constructor.