Add fences
Fences are now drawn on layer 2, represented by the new Fence class.
Ranch now has a domain and a range property (both DomainRanges), with
setters and getters. When one is set, Ranch converts it to Intervals
using the new method Fence.intervalToFences(). Drawing takes place
in Ranch.drawFences(), which delegates to Fence.draw().

Note that since the two fence graphics (fence_horizontal.png and
fence_vertical.png) are not really analogous in size, there's a lot
of finicky twiddling with the sizing going on in Fence.draw() and the
methods to which it delegates.

layer2 has been renamed to layerFences.