Make domain/range text boxes functional, style
The fences are now updated when the user presses enter in the domain or
range text boxes.

Also there is now decorative styled text around the domain/range boxes,
and they are styled to have a blue border usually, but a red border on
error (signaled through the "error" pseudo-class).