Add Gumdrop Joe in the bottom-left corner
Add GumdropJoe, a class which encapsulates the Gumdrop Joe drawing in
the bottom-left of the screen (stored in gumdrop_joe.png). GumdropJoe
not only draws the image, but also has support for a speech bubble
with text that Gumdrop Joe is saying. Ranch has an instance of
GumdropJoe, and it's drawn on a new layer (layerGumdropJoe).

Ranch has two new methods: gumdropJoeSay() and gumdropJoeClear(), which
simply update the text Gumdrop Joe is saying and then redraws Joe.