Add Ranch.distributeCows()
This new method distributes cows over two Intervals. It places cows in
the top-left/bottom-right or top-right/bottom-left corners (to guarantee
the size of the bounding box of the cows), then randomly distributes
cows across the rest of the bounding box. The number of cows is
determined by the area of the bounding box.

Also make Interval.EqualBoundsException a RuntimeException, change
Level0 to use Ranch.distributeCows() and be challenging, add
a DomainRange constructor that uses an array of Intervals.