Add play again button with placeholder image
There's now a play again button, encapsulated in PlayAgainButton. It
just draws a (currently placeholder) image at the bottom of the screen
if it's active. It's activated by its activate() method, and remains
active until it's clicked. It also implements EventHandler<MouseEvent>,
so it must be registered as such in order to handle clicks. It takes
its onClick callback as a constructor parameter.

Ranch fully encapsulates PlayAgainButton, drawing it on layerGumdropJoe.

There's also a new non-closeable Gumdrop Joe poem for when the user has
completed all the levels.