Add win image (animated!), real play again button
play_again.png is now an actual image instead of a pink rectangle.

win.png now contains the win image, and Controller animates it appearing
on victory. It appears (about) in the middle of the window, and to do
this main.fxml now has a StackLayout as the root element.